8253343: Extract G1 Young GC algorithm related code from G1CollectedHeap
